device_bridge/requirements.txt
websockets
pyautogui
Pillow
google-generativeai
